voice_photoshop/requirements.txt
# Core Framework
photoshop-python-api>=1.0.0

# ASR (Automatic Speech Recognition)
openai-whisper>=20231117

# Async & Concurrency
asyncio-mqtt>=0.11.0
aiofiles>=23.0.0

# Template & Config
Jinja2>=3.1.0
PyYAML>=6.0

# UI (Optional)
PySide6>=6.5.0

# Monitoring & Logging
structlog>=23.0.0
prometheus-client>=0.17.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# LLM (Optional)
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0
